Optionally submit the email while uploading the crash report

+++ main/core/src/EBox/SysInfo/CGI/CrashReport.pm
@@ -24,24 +24,37 @@ use EBox::Gettext;
 use EBox::Validate;
 use EBox::Util::BugReport;
 
+use constant OC_CRASH_UPLOAD => '/usr/share/openchange/upload-crash-report.py';
 my $CRASH_DIR = '/var/crash';
+my $CRASHREPORT_SERVER_URL = 'http://crashreport.zentyal.org/report/';
 
 sub _process
 {
     my ($self) = @_;
 
     my $action = $self->param('action');
+    my $email = $self->param('email');
+
+    if ($email) {
+        $email = "-n $email";
+    } else {
+        $email = ""
+    }
 
     # FIXME: unhardcode samba if more daemon crashes are watched
 
     if ($action eq 'report') {
         my @files = @{EBox::Sudo::root("ls $CRASH_DIR | grep ^_usr_sbin_samba")};
-        foreach my $file (@files) {
-            chomp($file);
-            EBox::info("Sending crash report: $file");
-            EBox::Sudo::root("/usr/share/zentyal/crash-report $CRASH_DIR/$file");
+        if (EBox::Sudo::fileTest('-x', OC_CRASH_UPLOAD)) {
+            foreach my $file (@files) {
+                chomp($file);
+                EBox::info("Sending crash report: $file");
+                EBox::Sudo::root(OC_CRASH_UPLOAD . " $email $CRASH_DIR/$file $CRASHREPORT_SERVER_URL");
+            }
+            EBox::Sudo::root('rm -f /var/crash/_usr_sbin_samba*');
+        } else {
+            EBox::error('Cannot send crash report as ' . OC_CRASH_UPLOAD . ' is not available');
         }
-        EBox::Sudo::root('rm -f /var/crash/_usr_sbin_samba*');
     } elsif ($action eq 'discard') {
         EBox::Sudo::root('rm -f /var/crash/_usr_sbin_samba*');
     }
